requirements.txt
google-adk
fastapi
uvicorn[standard]
cloudscraper
python-dotenv
requests
